fix: complete template component props
