+++ crates/vize_canon/src/virtual_ts/scope/mod.rs
@@ -8,6 +8,7 @@ mod closures;
 mod component_events;
 mod component_prop_checker;
 mod component_prop_expressions;
+mod component_prop_navigation;
 mod component_props;
 mod context;
 mod emit;
