+++ crates/vize_maestro/src/ide/completion.rs
@@ -18,6 +18,8 @@ mod style;
 pub(crate) mod template;
 
 #[cfg(test)]
+mod component_props_tests;
+#[cfg(test)]
 mod tests;
 
 // Cross-module reuse: inlay-hint code resolves reactive binding types with
